refactor game_display.fake_units_, a t_umap instead of a deque
this refactor is a first step to help with solving the double-kill bug,
as it reduces balkanization in the codebase. currently "real" units are
stored in and owned by the unit_map, and "fake" units are owned by the
game_display object. because units are large and balky, code that
needs to run fast e.g. the animation code, needs to work with pointers
rather than storing local copies.

currently we pass naked pointers to the animation code from either the
game_display or the unit_map, but this is unsafe, and can and does lead
to segfaults.

we can and should be passing some kind of smart pointer. there are
several options for what the unit_animation should be holding on to:

1) a pointer to a unit_pod
2) a unit_map::iterator
3) change unit map to hold shared_ptr's instead of unit_pods, and
make the animations hold either a shared_ptr or a weak_ptr
(debateable)

no matter what solution we end up with, this refactor is a sensible
first step, as now we wont have a separate code path for fake units
and real units in the animation code.